requirements.txt
# Django~=3.2.10
absl-py==2.0.0
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.9.3
billiard==4.2.0
bs4==0.0.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.4
cssselect==1.2.0
defusedxml==0.8.0rc2
Django==4.2.5
djangorestframework==3.14.0
dnspython==1.16.0
exceptiongroup==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.1
gast==0.5.4
gnews==0.3.6
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
jieba3k==0.35.1
joblib==1.3.2
jsonfield==3.1.0
keras==2.15.0
kombu==5.3.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
# nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
selenium==4.16.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
# tensorflow==2.15.0
# tensorflow-estimator==2.15.0
# tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
webdriver-manager==4.0.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
youtube-transcript-api==0.6.1